1. #0002827: При удалении групп раскраски файлов размер списка групп не уменьшается 2. Атрибуты в панелях, диалогах, результатах поиска и меню фильтров теперь в одном и том же порядке. 3. Неадекватно работала настройка колонок в результатах поиска. 4. Рефакторинг.

+++ far/strmix.cpp
@@ -774,7 +774,7 @@ string & FileSizeToStr(string &strDestStr, unsigned __int64 Size, int Width, uns
 // Return - количество замен
 int ReplaceStrings(string &strStr, const string& FindStr, const string& ReplStr, bool IgnoreCase, int Count)
 {
-	const int LenFindStr = static_cast<int>(FindStr.size());
+	const size_t LenFindStr = FindStr.size();
 	if ( !LenFindStr || !Count )
 		return 0;
